Merge 2.4 into develop

+++ packages/core-blockchain/src/state-machine.ts
@@ -3,18 +3,16 @@
 import { app } from "@arkecosystem/core-container";
 import { EventEmitter, Logger } from "@arkecosystem/core-interfaces";
 
-import { roundCalculator } from "@arkecosystem/core-utils";
-import { isException, models } from "@arkecosystem/crypto";
+import { isBlockChained, roundCalculator } from "@arkecosystem/core-utils";
+import { Blocks, Interfaces, Utils } from "@arkecosystem/crypto";
 
 import pluralize from "pluralize";
-import { config as localConfig } from "./config";
 import { blockchainMachine } from "./machines/blockchain";
 import { stateStorage } from "./state-storage";
-import { isBlockChained } from "./utils";
 
 import { Blockchain } from "./blockchain";
 
-const { Block } = models;
+const { BlockFactory } = Blocks;
 const config = app.getConfig();
 const emitter = app.resolvePlugin<EventEmitter.EventEmitter>("event-emitter");
 const logger = app.resolvePlugin<Logger.ILogger>("logger");
@@ -65,7 +63,7 @@ blockchainMachine.actionMap = (blockchain: Blockchain) => ({
             if (stateStorage.p2pUpdateCounter + 1 > 3) {
                 logger.info("Network keeps missing blocks.");
 
-                const networkStatus = await blockchain.p2p.checkNetworkHealth();
+                const networkStatus = await blockchain.p2p.getMonitor().checkNetworkHealth();
                 if (networkStatus.forked) {
                     stateStorage.numberOfBlocksToRollback = networkStatus.blocksToRollback;
                     event = "FORK";
@@ -125,30 +123,12 @@ blockchainMachine.actionMap = (blockchain: Blockchain) => ({
 
     async init() {
         try {
-            let block = await blockchain.database.getLastBlock();
-
-            if (!block) {
-                logger.warn("No block found in database");
-
-                block = new Block(config.get("genesisBlock"));
-
-                if (block.data.payloadHash !== config.get("network.nethash")) {
-                    logger.error("FATAL: The genesis block payload hash is different from configured the nethash");
-
-                    return blockchain.dispatch("FAILURE");
-                }
-
-                await blockchain.database.saveBlock(block);
-            }
+            const block: Interfaces.IBlock = await blockchain.database.getLastBlock();
 
             if (!blockchain.database.restoredDatabaseIntegrity) {
                 logger.info("Verifying database integrity");
 
-                const blockchainAudit = await blockchain.database.verifyBlockchain();
-                if (!blockchainAudit.valid) {
-                    logger.error("FATAL: The database is corrupted");
-                    logger.error(JSON.stringify(blockchainAudit.errors, null, 4));
-
+                if (!(await blockchain.database.verifyBlockchain())) {
                     return blockchain.dispatch("ROLLBACK");
                 }
 
@@ -159,6 +139,12 @@ blockchainMachine.actionMap = (blockchain: Blockchain) => ({
 
             // only genesis block? special case of first round needs to be dealt with
             if (block.data.height === 1) {
+                if (block.data.payloadHash !== config.get("network.nethash")) {
+                    logger.error("FATAL: The genesis block payload hash is different from configured the nethash");
+
+                    return blockchain.dispatch("FAILURE");
+                }
+
                 await blockchain.database.deleteRound(1);
             }
 
@@ -190,7 +176,7 @@ blockchainMachine.actionMap = (blockchain: Blockchain) => ({
             if (process.env.NODE_ENV === "test") {
                 logger.verbose("TEST SUITE DETECTED! SYNCING WALLETS AND STARTING IMMEDIATELY.");
 
-                stateStorage.setLastBlock(new Block(config.get("genesisBlock")));
+                stateStorage.setLastBlock(BlockFactory.fromJson(config.get("genesisBlock")));
                 await blockchain.database.buildWallets();
 
                 return blockchain.dispatch("STARTED");
@@ -202,12 +188,7 @@ blockchainMachine.actionMap = (blockchain: Blockchain) => ({
              * database init                 *
              ******************************* */
             // Integrity Verification
-            const verifiedWalletsIntegrity = await blockchain.database.buildWallets();
-            if (!verifiedWalletsIntegrity && block.data.height > 1) {
-                logger.warn(
-                    "Rebuilding wallets table because of some inconsistencies. Most likely due to an unfortunate shutdown.",
-                );
-            }
+            await blockchain.database.buildWallets();
 
             await blockchain.database.restoreCurrentRound(block.data.height);
             await blockchain.transactionPool.buildWallets();
@@ -221,8 +202,10 @@ blockchainMachine.actionMap = (blockchain: Blockchain) => ({
     },
 
     async downloadBlocks() {
-        const lastDownloadedBlock = stateStorage.lastDownloadedBlock || stateStorage.getLastBlock();
-        const blocks = await blockchain.p2p.downloadBlocks(lastDownloadedBlock.data.height);
+        const lastDownloadedBlock: Interfaces.IBlock = stateStorage.lastDownloadedBlock || stateStorage.getLastBlock();
+        const blocks: Interfaces.IBlockData[] = await blockchain.p2p
+            .getMonitor()
+            .syncWithNetwork(lastDownloadedBlock.data.height);
 
         if (blockchain.isStopped) {
             return;
@@ -233,8 +216,9 @@ blockchainMachine.actionMap = (blockchain: Blockchain) => ({
             return;
         }
 
-        const empty = !blocks || blocks.length === 0;
-        const chained = !empty && (isBlockChained(lastDownloadedBlock, { data: blocks[0] }) || isException(blocks[0]));
+        const empty: boolean = !blocks || blocks.length === 0;
+        const chained: boolean =
+            !empty && (isBlockChained(lastDownloadedBlock.data, blocks[0]) || Utils.isException(blocks[0]));
 
         if (chained) {
             logger.info(
@@ -256,7 +240,9 @@ blockchainMachine.actionMap = (blockchain: Blockchain) => ({
                 blockchain.dispatch("DOWNLOADED");
             } catch (error) {
                 logger.warn(`Failed to enqueue downloaded block.`);
+
                 blockchain.dispatch("NOBLOCK");
+
                 return;
             }
         } else {
@@ -265,6 +251,7 @@ blockchainMachine.actionMap = (blockchain: Blockchain) => ({
             } else {
                 logger.warn(`Downloaded block not accepted: ${JSON.stringify(blocks[0])}`);
                 logger.warn(`Last downloaded block: ${JSON.stringify(lastDownloadedBlock.data)}`);
+
                 blockchain.clearQueue();
             }
 
@@ -283,11 +270,12 @@ blockchainMachine.actionMap = (blockchain: Blockchain) => ({
 
     async startForkRecovery() {
         logger.info("Starting fork recovery");
+
         blockchain.clearAndStopQueue();
 
         await blockchain.database.commitQueuedQueries();
 
-        const random = 4 + Math.floor(Math.random() * 99); // random int inside [4, 102] range
+        const random: number = 4 + Math.floor(Math.random() * 99); // random int inside [4, 102] range
 
         await blockchain.removeBlocks(stateStorage.numberOfBlocksToRollback || random);
         stateStorage.numberOfBlocksToRollback = null;
@@ -295,7 +283,7 @@ blockchainMachine.actionMap = (blockchain: Blockchain) => ({
         logger.info(`Removed ${pluralize("block", random, true)}`);
 
         await blockchain.transactionPool.buildWallets();
-        await blockchain.p2p.refreshPeersAfterFork();
+        await blockchain.p2p.getMonitor().refreshPeersAfterFork();
 
         blockchain.dispatch("SUCCESS");
     },
@@ -303,29 +291,26 @@ blockchainMachine.actionMap = (blockchain: Blockchain) => ({
     async rollbackDatabase() {
         logger.info("Trying to restore database integrity");
 
-        const { maxBlockRewind, steps } = localConfig.get("databaseRollback");
-        let blockchainAudit;
+        const { maxBlockRewind, steps } = app.resolveOptions("p2p").databaseRollback;
 
         for (let i = maxBlockRewind; i >= 0; i -= steps) {
             await blockchain.removeTopBlocks(steps);
 
-            blockchainAudit = await blockchain.database.verifyBlockchain();
-            if (blockchainAudit.valid) {
+            if (await blockchain.database.verifyBlockchain()) {
                 break;
             }
         }
 
-        if (!blockchainAudit.valid) {
+        if (!(await blockchain.database.verifyBlockchain())) {
             // TODO: multiple attempts? rewind further? restore snapshot?
-            logger.error("FATAL: Failed to restore database integrity");
-            logger.error(JSON.stringify(blockchainAudit.errors, null, 4));
             blockchain.dispatch("FAILURE");
+
             return;
         }
 
         blockchain.database.restoredDatabaseIntegrity = true;
 
-        const lastBlock = await blockchain.database.getLastBlock();
+        const lastBlock: Interfaces.IBlock = await blockchain.database.getLastBlock();
         logger.info(
             `Database integrity verified again after rollback to height ${lastBlock.data.height.toLocaleString()}`,
         );
